# Add benchmark for simple indexing + some simple optimisations 