--- a/docs/changelogs/vNEXT.rst
+++ b/docs/changelogs/vNEXT.rst
@@ -1,24 +1,8 @@
 =====
-NEXT
+0.5.2
 =====
 
 **Release date:** XXXX-XX-XX
-
-Breaking changes
-================
-
-.. TODO remove warning and replace with "None" if no breaking
-   changes.
-
-.. warning:: This release contains breaking changes. Be sure
-   to follow migration steps before upgrading.
-
-Breaking change 1
------------------
-
-- summary of breaking change
-
-Link to migration guide
 
 
 New features
@@ -37,7 +21,5 @@ Link to relevant documentation section
 Other bugs & improvements
 =========================
 
-- bulleted list
-- improvements
-- of smaller
-- potentially with doc links
+- Improve rule indexing performance
+